requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
altair==5.4.1
annotated-types==0.7.0
anyio==4.6.2.post1
attrs==24.2.0
backoff==2.2.1
beautifulsoup4==4.12.3
blinker==1.9.0
boto3==1.35.59
botocore==1.35.59
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.0
click==8.1.7
cryptography==43.0.3
dataclasses-json==0.6.7
dnspython==2.7.0
emoji==2.14.0
eval_type_backport==0.2.0
fastapi==0.115.5
filetype==1.2.0
frozenlist==1.5.0
gitdb==4.0.11
GitPython==3.1.43
greenlet==3.1.1
h11==0.14.0
html5lib==1.1
httpcore==1.0.6
httpx==0.27.2
httpx-sse==0.4.0
idna==3.10
iniconfig==2.0.0
Jinja2==3.1.4
jmespath==1.0.1
joblib==1.4.2
jsonpatch==1.33
jsonpath-python==1.0.6
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
langchain==0.3.7
langchain-aws==0.2.7
langchain-community==0.3.7
langchain-core==0.3.17
langchain-text-splitters==0.3.2
langdetect==1.0.9
langsmith==0.1.142
lxml==5.3.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.23.1
mdurl==0.1.2
mongoengine==0.29.1
multidict==6.1.0
mypy-extensions==1.0.0
narwhals==1.13.4
nest-asyncio==1.6.0
nltk==3.9.1
numpy==1.26.4
olefile==0.47
orjson==3.10.11
packaging==24.2
pandas==2.2.3
pillow==11.0.0
plotly==5.24.1
pluggy==1.5.0
propcache==0.2.0
protobuf==5.28.3
psutil==6.1.0
pyarrow==18.0.0
pycparser==2.22
pydantic==2.9.2
pydantic-settings==2.6.1
pydantic_core==2.23.4
pydeck==0.9.1
Pygments==2.18.0
pymongo==4.10.1
pypdf==5.1.0
pytest==8.3.3
python-dateutil==2.8.2
python-dotenv==1.0.1
python-iso639==2024.10.22
python-magic==0.4.27
python-oxmsg==0.0.1
pytz==2024.2
PyYAML==6.0.2
RapidFuzz==3.10.1
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
requests-mock==1.12.1
requests-toolbelt==1.0.0
rich==13.9.4
rpds-py==0.21.0
s3transfer==0.10.3
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.35
starlette==0.41.2
streamlit==1.40.1
tenacity==9.0.0
tiktoken==0.8.0
toml==0.10.2
tornado==6.4.1
tqdm==4.67.0
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
unstructured==0.16.5
unstructured-client==0.27.0
urllib3==2.2.3
webencodings==0.5.1
wrapt==1.16.0
yarl==1.17.1
